requirements.txt
Babel==2.6.0
Django==2.1.1
django-countries==5.3.2
django-extensions==2.1.2
django-phonenumber-field==2.0.1
djangorestframework==3.8.2
phonenumbers==8.9.12
pytz==2018.5
six==1.11.0
